Phew..... ok refactored the CPU to use stack pointer for branching to and returning from subroutines.
- Updated to 7-bit CPU cycles so can accomodate the 1 (!) extra cycle needed to
push both bytes of the PC to the stack and do a branch (5 cycles needed).
This means that the EEPROM for CPU controller now needs to be 1024 bytes instead of 512 bytes :(

- Added an END microcode instruction so dont have to go through all 8 CPU cycles each time,
and each instruction only takes as long as required to do its thing.

Updated the spreadsheet and makerom.py

Todo: update assembler and emulator to handle the new addressing modes and stack